backend/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
